requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# LLM and AI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
openai>=1.3.0
google-generativeai>=0.3.0

# Vector Database
qdrant-client>=1.6.0

# PDF Processing
PyPDF2>=3.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0

# Optional for enhanced functionality
tiktoken>=0.5.0
faiss-cpu>=1.7.4
